requirements.txt
asgiref==3.5.0
Django==4.0.2
django-guardian==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
